RTE_PKTMBUF_HEADROOM needs to be changed to 256 from 128.
During the dpdk upgrade from 18.05 to 19.11, the size of RTE_PKTMBUF_HEADROOM was reverted from
256 to 128 with the assumption that the mbuff for mirror metadata which needed the headroom of
size greater than 128 was now being chained so the RTE_PKTMBUF_HEADROOM could be reverted back
to 128. But it turns out other packets can have a header length of size greater than 128 too.
For instance, an IPv6 TCP packet with VXLAN header has a header of size 140 and this packet
was getting dropped.

Fix: Changing the size of RTE_PKTMBUF_HEADROOM back to 256 to accomodate headrooms for different
kinds of encapsulated packets.

Partial-Jira-Bug: CEM-23810
Signed-off-by: Vinay Agrawal <vagrawal@juniper.net>
Change-Id: Ie17a05b082dcb88e1cdb58c8040301cf9607f028