CEM-26032: vRouter dpdk crashing in vr_ip_csum function.
Issue Description: vRouter dpk started crashing on sanity tests for mirror interface test cases when the packet hit the vr_ip_csum() function.

RCA: When RTE_PKTMBUF_HEADROOM was changed from 128 to 256 for performance improvements as a part of CEM-23810 for gso packets with headroom > 128, it caused a breakage during access and operation on the pointer pointing to the mbuf for the mirroring packet in the vr_ip_csum() during checksum calculation for mirroring scenario.

Fix: Reverting the RTE_PKTMBUF_HEADROOM from 256 back to 128 will fix this issue and since CEM-23810 was a rss_mempool memory leak issue, it was fixed as a part of https://gerrit.tungsten.io/r/c/tungstenfabric/tf-vrouter/+/67230.

Note: We will continue to debug why the mirroring scenario causes a crash with HEADROOM 256 but not with 128 on the same codepath and try to add another logic to make performance improvements for gso packets with large headroom which were discovered in CEM-23810.

Closes-Jira-Bug: CEM-26032
Signed-off-by: Vinay Agrawal <vagrawal@juniper.net>
Change-Id: Ie80609efb8e1770fd7f6712af15987372e1d1101